# Add optimized get_darkest_area implementation